[Cart] Rename cart query
