+++ spec/ViewRepository/CartViewRepositorySpec.php
@@ -2,17 +2,17 @@
 
 declare(strict_types = 1);
 
-namespace spec\Sylius\ShopApiPlugin\Query;
+namespace spec\Sylius\ShopApiPlugin\ViewRepository;
 
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Repository\OrderRepositoryInterface;
 use Sylius\ShopApiPlugin\Factory\CartViewFactoryInterface;
-use Sylius\ShopApiPlugin\Query\CartQueryInterface;
+use Sylius\ShopApiPlugin\ViewRepository\CartViewRepositoryInterface;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Sylius\ShopApiPlugin\View\CartSummaryView;
 
-final class CartQuerySpec extends ObjectBehavior
+final class CartViewRepositorySpec extends ObjectBehavior
 {
     function let(OrderRepositoryInterface $cartRepository, CartViewFactoryInterface $cartViewFactory)
     {
@@ -21,21 +21,20 @@ final class CartQuerySpec extends ObjectBehavior
 
     function it_is_cart_query()
     {
-        $this->shouldImplement(CartQueryInterface::class);
+        $this->shouldImplement(CartViewRepositoryInterface::class);
     }
 
-    function it_is_provide_cart_view(
+    function it_provides_cart_view(
         OrderRepositoryInterface $cartRepository,
         CartViewFactoryInterface $cartViewFactory,
         OrderInterface $cart,
         CartSummaryView $cartView
-    )
-    {
+    ) {
         $cartRepository->findOneBy(['tokenValue' => 'ORDERTOKEN'])->willReturn($cart);
         $cart->getLocaleCode()->willReturn('en_GB');
 
         $cartViewFactory->create($cart, 'en_GB')->willReturn($cartView);
 
-        $this->findByToken('ORDERTOKEN')->shouldReturn($cartView);
+        $this->getOneByToken('ORDERTOKEN')->shouldReturn($cartView);
     }
 }
