+++ src/Resources/config/services/actions/cart.xml
@@ -5,7 +5,7 @@
         <service id="sylius.shop_api_plugin.controller.cart.summarize_action"
                  class="Sylius\ShopApiPlugin\Controller\Cart\SummarizeAction"
         >
-            <argument type="service" id="sylius.shop_api_plugin.query.cart_query" />
+            <argument type="service" id="sylius.shop_api_plugin.view_repository.cart_view_repository" />
             <argument type="service" id="fos_rest.view_handler" />
         </service>
 
@@ -52,7 +52,7 @@
             <argument type="service" id="tactician.commandbus" />
             <argument type="service" id="validator" />
             <argument type="service" id="sylius.shop_api_plugin.factory.validation_error_view_factory" />
-            <argument type="service" id="sylius.shop_api_plugin.query.cart_query" />
+            <argument type="service" id="sylius.shop_api_plugin.view_repository.cart_view_repository" />
         </service>
     </services>
 </container>
