+++ src/Resources/config/services/queries.xml
@@ -2,7 +2,7 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="sylius.shop_api_plugin.query.cart_query" class="Sylius\ShopApiPlugin\Query\CartQuery">
+        <service id="sylius.shop_api_plugin.view_repository.cart_view_repository" class="Sylius\ShopApiPlugin\ViewRepository\CartViewRepository">
             <argument type="service" id="sylius.repository.order" />
             <argument type="service" id="sylius.shop_api_plugin.factory.cart_view_factory" />
         </service>
